Move tests in new package
